+++ src/main/java/com/example/controller/RegionController.java
@@ -1,10 +1,15 @@
 package com.example.controller;
 
 import com.example.dto.ArticleTypeDTO;
+import com.example.dto.JWTDTO;
 import com.example.dto.RegionDTO;
 import com.example.entity.RegionEntity;
+import com.example.enums.LangEnum;
+import com.example.enums.ProfileRole;
 import com.example.service.RegionService;
+import com.example.util.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,24 +22,39 @@ public class RegionController {
     @Autowired
     private RegionService regionService;
     @PostMapping("/create")
-    public ResponseEntity<RegionDTO> create(@RequestBody RegionDTO dto){
+    public ResponseEntity<RegionDTO> create(@RequestBody RegionDTO dto,
+                                            @RequestHeader(value = "Authorization")String jwt){
+        JWTDTO jwtdto= JWTUtil.decode(jwt);
+        if (!jwtdto.getRole().equals(ProfileRole.ADMIN)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(regionService.create(dto));
     }
     @PutMapping("/update/{id}")
-    public ResponseEntity<Boolean>update(@PathVariable("id")Integer id ,@RequestBody RegionDTO dto){
+    public ResponseEntity<Boolean>update(@PathVariable("id")Integer id ,@RequestBody RegionDTO dto,
+                                         @RequestHeader(value = "Authorization")String jwt){
+        JWTDTO jwtdto= JWTUtil.decode(jwt);
+        if (!jwtdto.getRole().equals(ProfileRole.ADMIN)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(regionService.update(id,dto));
     }
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Boolean>deleteById(@PathVariable("id") Integer id){
+    public ResponseEntity<Boolean>deleteById(@PathVariable("id") Integer id,
+                                             @RequestHeader(value = "Authorization")String jwt){
+        JWTDTO jwtdto= JWTUtil.decode(jwt);
+        if (!jwtdto.getRole().equals(ProfileRole.ADMIN)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(regionService.delete(id));
     }
     @GetMapping("/all")
     public ResponseEntity<List<RegionDTO>>getAll(){
         return ResponseEntity.ok(regionService.getAll());
     }
-    @GetMapping("/lang/{id}/{language}")
-    public ResponseEntity<Optional<RegionDTO>> getArticleTypesByLanguage(@PathVariable Integer id, @PathVariable String language) {
-        Optional<RegionDTO> region =regionService.getByLang(id,language);
+    @GetMapping("/lang/{language}")
+    public ResponseEntity<List<RegionDTO>> getArticleTypesByLanguage(@PathVariable LangEnum language) {
+        List<RegionDTO> region =regionService.getByLang(language);
         return ResponseEntity.ok(region);
     }
 }
